+++ src/client/src/graphql/mutations/createBaseInvoice.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const CREATE_BASE_INVOICE = gql`
-  mutation CreateBaseInvoice($amount: Int!) {
+  mutation CreateBaseInvoice($amount: Float!) {
     createBaseInvoice(amount: $amount) {
       request
       id
