+++ src/client/src/graphql/mutations/keysend.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const KEY_SEND = gql`
-  mutation Keysend($destination: String!, $tokens: Int!) {
+  mutation Keysend($destination: String!, $tokens: Float!) {
     keysend(destination: $destination, tokens: $tokens) {
       is_confirmed
     }
