+++ src/client/src/graphql/queries/getForwards.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const GET_FORWARDS = gql`
-  query GetForwards($days: Int!) {
+  query GetForwards($days: Float!) {
     getForwards(days: $days) {
       created_at
       fee
